refactor(apps,website): Use the same `generate_schema_json()` function (#18025)
The initial goal was to extract the schema generation utility from `oxfmtrc.rs`.

However, I found that `oxc_linter` also defines a similar function, so I unified them into `website_common`.

- Added `generate_schema_json<T>()` to `website_common`
- Removed duplicated implementations from `oxfmtrc.rs` and `oxlintrc.rs`
- Moved `test_schema_json` from `oxc_linter` to `website_linter` too

NOTE: When running `cargo t -p oxc_linter`, it previously generated `npm/oxlint/configuration_schema.json`. Now use `just linter-schema-json` instead.
