+++ tasks/website_linter/src/json_schema.rs
@@ -1,10 +1,26 @@
 use oxc_linter::Oxlintrc;
 use schemars::schema_for;
-use website_common::Renderer;
+use website_common::{Renderer, generate_schema_json};
 
 #[expect(clippy::print_stdout)]
 pub fn print_schema_json() {
-    println!("{}", Oxlintrc::generate_schema_json());
+    println!("{}", generate_schema_json::<Oxlintrc>());
+}
+
+#[test]
+fn test_schema_json() {
+    use project_root::get_project_root;
+    use std::fs;
+
+    let path = get_project_root().unwrap().join("npm/oxlint/configuration_schema.json");
+    let json = generate_schema_json::<Oxlintrc>();
+    let existing_json = fs::read_to_string(&path).unwrap_or_default();
+    if existing_json.trim() != json.trim() {
+        fs::write(&path, &json).unwrap();
+    }
+    insta::with_settings!({ prepend_module_to_snapshot => false }, {
+        insta::assert_snapshot!(json);
+    });
 }
 
 #[test]
